Article的id类型改为Long

+++ src/main/java/com/kkkoke/networkrepair/pojo/Article.java
@@ -15,7 +15,7 @@ import javax.persistence.Id;
 public class Article {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;//文章id
+    private Long id;//文章id
     private String createTime;//文章创建时间
     private String updateTime;//文章修改时间
     private String contentPath;//文章内容路径
